refactor(cli): split main.rs into a module structure
main.rs had grown to 1555 lines mixing the clap CLI surface, per-
subcommand orchestration, and config-merging logic. Split into
cli.rs (arg/subcommand definitions), config_merge.rs (CLI-flag/config
reconciliation), commands/{convert,parse,index,init}.rs (subcommand
handlers), and tests.rs, leaving main.rs as a thin entry point. No
behavior change -- --help output is byte-identical before and after.
